Report an error when `UNKNOWN_TAG` is found #55
The cause of the error was the method ​`p_unknown_tag`
in `parsers/tagvalue.py`. Due to the incorrect context-free
grammar specification defined in the method, the line after
the `unknown_tag` was not taken into consideration. The
context-free grammar specification was rectified and tests
were corresponding tests were added.

Signed-off-by: Yash Nisar <yash.nisar@somaiya.edu>