Report correct line no.s while showing errors #77
This occurs due to the method `t_text_end` in the
file `lexers\tagvalue.py`. The line `t.value.count('\n')`
doesn't detect newlines because the earlier line is stripped
of whitespaces and newlines. This issue is resolved by first
counting the no. of newline chars and then stripping
whitespaces and newlines so information is preserved.

Signed-off-by: Yash Nisar <yash.nisar@somaiya.edu>