Return messages instead of modifying list #50
We now return messages from each `validation` function
instead of modifying lists in place. It is preferrable to
have side-effect free functions, so in this case, return a
new list with the element modified. There are a few rare
corner use cases where in place changes can make sense,
but 99% of the times, it a bad idea and can trigger any
kind of weird, byzantine and hard to debug bugs.

Signed-off-by: Yash Nisar <yash.nisar@somaiya.edu>