examples:write_tv: add friendlier error display
The current version of the example fails with "Document is Invalid"
errors, which makes it rather hard to understand what is invalid about
the document. Further, the  exception caught itself presents is already
floating information about which checks specifically made the document
invalid. Present this information to the user in such a way that they
can identify and fix the errors in their document