Tie click version for Python 2.x.
Click dropped support for all 2.x versions and all 3.5 or earlier
versions of Python.  So, in order to support those versions, we
must tie the click version to click 7.x.

Since we're still actively testing these older Python versions
in CI, this should fix a number of issues in CI.

Fixes #161.

Signed-off-by: Jeff Licquia <licquia@linuxfoundation.org>