[issue-479] fix parsing of snippet ranges
The previous regular expressions for ranges would prevent lines with ":" to be parsed as LINE and instead be parsed as RANGE, the remainder after ":\d+" would be truncated and ignored, this lead to problems when parsing a package version like "1:2.36.1-8+deb11u1". Instead of defining a separate lexer function for RANGE we check if a LINE token matches the RANGE-pattern. This allows LINEs to include the RANGE pattern without getting the token RANGE.

Signed-off-by: Meret Behrens <meret.behrens@tngtech.com>