Replace typeguard type checking with beartype.
typeguard was prohibitively slow on large inputs, beartype is
significantly faster. However, beartype comes with a caveat in that it
does not do deep nested type checks, but instead utilizes randomized
sampled testing. As such, this still allows certain misuses of types to
go undetected and may not be our final solution. However, in the
meantime we deem it is much better to have beartype than not have any
type checking taking place.

Signed-off-by: Holger Frydrych <holger.frydrych@tngtech.com>