Execute test_graph_generation as part of the regular test run rather than a separate step in the pipeline.
Background: there is currently an issue with newer patch versions of
Python 3.7 on macos, which are missing bz2 support, thus leading to a
failure to include the optional networkx package. This in turn results
in that particular test being skipped. But if it runs as a separate step
in the pipeline, this means that no tests will execute at all in that
step and it will then be marked as an overall failure.

See https://github.com/actions/setup-python/issues/682

Signed-off-by: Holger Frydrych <holger.frydrych@tngtech.com>